Move article-inputs/ to common/
