+++ components/cartes-impact/simple-card/impact-impots.jsx
@@ -4,7 +4,7 @@ import Typography from "@material-ui/core/Typography";
 import PropTypes from "prop-types";
 import { Fragment, PureComponent } from "react";
 
-import { Parameter } from "../../articles-inputs/parameter";
+import { Parameter } from "../../common";
 import { formatNumber } from "../../utils";
 import styles2 from "./impact-impots.module.scss";
