+++ components/common/articles-inputs/base-input-output/base-input-output-component.jsx
