+++ components/common/articles-inputs/formula-output/formula-output-component.jsx
