+++ components/common/articles-inputs/index.js
@@ -1,3 +1,3 @@
-export { Parameter } from "./parameter";
+export { Parameter, ResultValues, StateParameter } from "./parameter";
 export { default as BaseInputOutput } from "./base-input-output";
 export { default as FormulaOutput } from "./formula-output";
