+++ components/common/index.ts
@@ -1,3 +1,4 @@
 export { default as AppHeader } from "./app-header";
+export * from "./articles-inputs";
 export * from "./card"
 export * from "./expandable-panels";
