+++ components/dotations/articles/dsr-fraction-bourg-centre/DsrFractionBourgCentre.tsx
@@ -1,6 +1,6 @@
 import { Fragment, PureComponent } from "react";
 
-import { StateParameter } from "../../../articles-inputs/parameter";
+import { StateParameter } from "../../../common";
 
 export class DsrFractionBourgCentre extends PureComponent {
   render() {
