+++ components/dotations/articles/dsr-fraction-cible/DsrFractionCible.tsx
@@ -1,6 +1,6 @@
 import { Fragment, PureComponent } from "react";
 
-import { StateParameter } from "../../../articles-inputs/parameter";
+import { StateParameter } from "../../../common";
 
 export class DsrFractionCible extends PureComponent {
   render() {
