+++ components/dotations/articles/montant-dgf/MontantDgf.tsx
@@ -1,6 +1,6 @@
 import { Fragment, PureComponent } from "react";
 
-import { StateParameter } from "../../../articles-inputs/parameter";
+import { StateParameter } from "../../../common";
 import { ExpandableText } from "../../../common";
 
 export class MontantDgf extends PureComponent {
