+++ components/dotations/results/commune-type/eligibilite/Eligibilite.tsx
@@ -6,7 +6,7 @@ import { connect, ConnectedProps } from "react-redux";
 // eslint-disable-next-line no-unused-vars
 import { RootState } from "../../../../../redux/reducers";
 import { getResultBoolValues } from "../../../../../redux/utils";
-import { Parameter } from "../../../../articles-inputs";
+import { Parameter } from "../../../../common";
 import styles from "./Eligibilite.module.scss";
 
 function getEligibilite(value: boolean|undefined): "éligible"|"non éligible"|undefined {
